requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
chromadb
pypdf
tiktoken
streamlit
pydantic
python-dotenv